// src/app/model/customer.model.ts
export interface Customer {
    _id: string;
    customerNumber: string;
    customerId?: any;
    firstName: string;
    lastName: string;
    fullName: string;
    email: string;
    emailOptIn: boolean;
    address: string;
    birthday: string;
    mobilePhone: string;
    homePhone: string;
    workPhone: string;
    spouseFullName: string;
    spouseBirthday: string;
    created: any;
    livevalue?: number;
  }